+++ ToyBox/classes/MonkeyPatchin/BagOfPatches/Limits/Unrestricted.cs
@@ -94,56 +94,11 @@ namespace ToyBox.BagOfPatches {
             }
         }
         
-        [HarmonyPatch(typeof(BlueprintSettlementBuilding), "CheckRestrictions", new Type[] { typeof(SettlementState) })]
-        public static class BlueprintSettlementBuilding_CheckRestrictions_Patch1 {
-            public static void Postfix(ref bool __result) {
-                if (settings.toggleSettlementRestrictions) {
-                    __result = true;
-                }
-            }
-        }
-
-        [HarmonyPatch(typeof(BlueprintSettlementBuilding), "CheckRestrictions", new Type[] { typeof(SettlementState), typeof(SettlementGridTopology.Slot) })]
-        public static class BlueprintSettlementBuilding_CheckRestrictions_Patch2 {
-            public static void Postfix(ref bool __result) {
-                if (settings.toggleSettlementRestrictions) {
-                    __result = true;
-                }
-            }
-        }
-
         [HarmonyPatch(typeof(Spellbook), nameof(Spellbook.CasterLevel), MethodType.Getter)]
         public static class Spellbook_CasterLevel_Patch {
             public static void Postfix(ref int __result, Spellbook __instance) {
                 if (settings.toggleUncappedCasterLevel) {
                     __result += __instance.m_BaseLevelInternal - __instance.BaseLevel - __instance.Blueprint.CasterLevelModifier;
-                }
-            }
-        }
-
-        [HarmonyPatch(typeof(SettlementState), "CanBuildUprgade", new Type[] { typeof(BlueprintSettlementBuilding) })]
-        public static class SettlementState_CanBuildUprgade_Patch {
-            public static void Postfix(ref bool __result) {
-                if (settings.toggleSettlementRestrictions) {
-                    __result = true;
-                }
-            }
-        }
-
-        [HarmonyPatch(typeof(SettlementState), "CanBuildByLevel", new Type[] { typeof(BlueprintSettlementBuilding) })]
-        public static class SettlementState_CanBuildByLevel_Patch {
-            private static void Postfix(ref bool __result) {
-                if (settings.toggleSettlementRestrictions) {
-                    __result = true;
-                }
-            }
-        }
-
-        [HarmonyPatch(typeof(SettlementState), "CanBuild", new Type[] { typeof(BlueprintSettlementBuilding) })]
-        public static class SettlementState_CanBuild_Patch {
-            public static void Postfix(ref bool __result) {
-                if (settings.toggleSettlementRestrictions) {
-                    __result = true;
                 }
             }
         }
